few more algorithms added

+++ Maximum Area of a Piece of Cake After Horizontal and Vertical Cuts/src/maximum/area/of/a/piece/of/cake/after/horizontal/and/vertical/cuts/MaximumAreaOfAPieceOfCakeAfterHorizontalAndVerticalCuts.java
@@ -74,7 +74,7 @@ public class MaximumAreaOfAPieceOfCakeAfterHorizontalAndVerticalCuts {
                 mv=verticalCuts[i]-verticalCuts[i-1];
             }
         }
-        return (int)((mh%1000000007*mv%1000000007));
+        return (int)(((long)mh%1000000007*mv%1000000007));
     }
     
     public static void main(String[] args) {
